Add `ProjectAnnotator`; Remove `ModelFileAnnotator`
